requirements.txt
flask~=2.3.3
flask_socketio
python-ffmpeg
werkzeug~=2.3.7
simple-websocket